+++ src/core/hash.rs
@@ -1,7 +1,25 @@
 use serde::{Deserialize, Serialize};
 use sha3::{Digest, Sha3_256};
+use std::fmt::Debug;
 
-use crate::core::Hash;
+use super::{AutoDeserialize, AutoHash, AutoSerialize, MemberBound};
+
+pub trait Hash: Debug + Clone + 'static {
+    /// The length in bytes of the Hasher output
+    const LENGTH: usize;
+
+    type Output: MemberBound
+        + AutoSerialize
+        + AutoDeserialize
+        + AutoHash
+        + AsRef<[u8]>
+        + AsMut<[u8]>
+        + Default
+        + Copy
+        + PartialOrd;
+
+    fn hash(s: &[u8]) -> Self::Output;
+}
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Sha3Hasher;
@@ -20,8 +38,8 @@ impl Hash for Sha3Hasher {
 
 #[cfg(test)]
 mod tests {
+    use crate::core::hash::Hash;
     use crate::core::hash::Sha3Hasher;
-    use crate::core::Hash;
 
     #[test]
     fn test_sha3_works() {
